# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для системы мозаичных купонов
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Mosaic API
  version: "1.0"
paths:
...
  /admin/partners/{id}/statistics:
    get:
      description: Returns detailed statistics for a specific partner
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Partner statistics
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Partner not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get partner statistics
      tags:
      - admin-partners
...
  /coupons:
    get:
      description: Returns list of coupons with filtering capabilities by code, status,
        size, style and partner
      parameters:
      - description: Coupon code for search
        in: query
        name: code
        type: string
      - description: Coupon status (new, used)
        in: query
        name: status
        type: string
      - description: Coupon size
        in: query
        name: size
        type: string
      - description: Coupon style
        in: query
        name: style
        type: string
      - description: Partner ID
        in: query
        name: partner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of coupons
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get coupons list with filtering
      tags:
      - coupons
...
  /coupons/{id}/download-materials:
    get:
      description: Downloads archive with redeemed coupon materials (original, preview,
        schema)
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: ZIP archive with materials
          schema:
            type: string
        "400":
          description: Invalid coupon ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Coupon not found or not used
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Download coupon materials
      tags:
      - coupons
...
  /coupons/export:
    get:
      description: Exports coupons to text file with filtering options
      parameters:
      - description: Partner ID for filtering
        in: query
        name: partner_id
        type: string
      - description: Coupon status for export
        in: query
        name: status
        type: string
      - description: 'Export format: codes (codes only) or full (full information)'
        in: query
        name: format
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Text file with coupons
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Export coupons
      tags:
      - coupons
...
  /coupons/statistics:
    get:
      description: Returns coupon statistics with optional partner filtering
      parameters:
      - description: Partner ID for filtering
        in: query
        name: partner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Coupon statistics
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get coupon statistics
      tags:
      - coupons
  /partner/coupons:
    get:
      description: Returns a list of coupons for the current partner with filtering
        and pagination
      parameters:
      - description: Coupon code for search
        in: query
        name: code
        type: string
      - description: Coupon status (new, used, completed)
        in: query
        name: status
        type: string
      - description: Coupon size
        in: query
        name: size
        type: string
      - description: Coupon style
        in: query
        name: style
        type: string
      - description: Creation date from (RFC3339 format)
        in: query
        name: created_from
        type: string
      - description: Creation date to (RFC3339 format)
        in: query
        name: created_to
        type: string
      - description: Usage date from (RFC3339 format)
        in: query
        name: used_from
        type: string
      - description: Usage date to (RFC3339 format)
        in: query
        name: used_to
        type: string
      - default: created_at
        description: Field to sort by
        enum:
        - created_at
        - updated_at
        - used_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of partner coupons with pagination info
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: JWT token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: User does not have partner role'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error when retrieving coupons
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get partner's coupons
      tags:
      - partner-coupons
  /partner/coupons/{id}/download-materials:
    get:
      description: Downloads archive with materials of a redeemed partner's coupon
        (original, preview, scheme)
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: ZIP archive with materials as attachment
          schema:
            type: string
        "400":
          description: 'Bad request: Invalid coupon ID format'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: JWT token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Coupon does not belong to partner'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Coupon not found or materials not available
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error during download process
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Download coupon materials
      tags:
      - partner-coupons
  /partner/coupons/export:
    get:
      description: Exports partner's coupons with status "new" in .txt or .csv format
      parameters:
      - default: csv
        description: File format (txt or csv)
        enum:
        - txt
        - csv
        in: query
        name: format
        type: string
      - default: new
        description: Coupon status to export (new, used, completed, or 'all' for all
          statuses)
        in: query
        name: status
        type: string
      produces:
      - text/plain
      - text/csv
      responses:
        "200":
          description: File with coupons as attachment
          schema:
            type: string
        "400":
          description: 'Bad request: Invalid parameters'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: JWT token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: User does not have partner role'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error during export process
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export partner's coupons
      tags:
      - partner-coupons
  /partner/dashboard:
    get:
      description: Returns data for the partner's main dashboard page including coupon
        statistics and recent activations
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard data with coupons and recent activations
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: JWT token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: User does not have partner role'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error when retrieving dashboard data
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get partner dashboard data
      tags:
      - partner-dashboard
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"